bot-config/requirements.txt
nose>=1.3.7
cement
mock
fire
paramiko
pyYaml